+++ EduTesting/EduTesting.Core/ViewModels/Test/TestResultRatingViewModel.cs
@@ -8,7 +8,7 @@ namespace EduTesting.ViewModels.Test
     public class TestResultRatingViewModel
     {
         public int RatingId { get; set; }
-        public int RatingLowerBound { get; set; }
+        public decimal RatingLowerBound { get; set; }
         public string RatingTitle { get; set; }
     }
 }
